requirements.txt
pefile
pillow